RoutineRequestDTO 추가

+++ src/main/java/com/example/youtube_routine/Entity/Routine.java
@@ -11,7 +11,7 @@ import java.time.LocalTime;
 public class Routine {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id; // 루틴 별 고유 id
 
     @Column(nullable = false)
     private Day day; // 요일
